+++ coordinates/coordinates_2d.hpp
@@ -218,7 +218,7 @@ namespace scifir
 
 			float get_p() const
 			{
-				return std::sqrt(std::pow(x,2) + std::pow(y,2));
+				return float(std::sqrt(std::pow(x,2) + std::pow(y,2)));
 			}
 
 			angle get_theta() const
@@ -266,7 +266,7 @@ namespace scifir
 
 			float distance_to_origin() const
 			{
-				return std::sqrt(std::pow(x,2) + std::pow(y,2));
+				return float(std::sqrt(std::pow(x,2) + std::pow(y,2)));
 			}
 
 			string display_polar() const
@@ -314,7 +314,7 @@ namespace scifir
 
 	inline float cartesian_2d_to_polar_r(float x,float y)
 	{
-		return std::sqrt(std::pow(x,2) + std::pow(y,2));
+		return float(std::sqrt(std::pow(x,2) + std::pow(y,2)));
 	}
 
 	inline float cartesian_2d_to_polar_theta(float x,float y)
