+++ coordinates/coordinates_3dr.hpp
@@ -566,7 +566,7 @@ namespace scifir
 
 			float get_p() const
 			{
-				return std::sqrt(std::pow(x,2) + std::pow(y,2));
+				return float(std::sqrt(std::pow(x,2) + std::pow(y,2)));
 			}
 
 			angle get_spherical_theta() const
@@ -576,7 +576,7 @@ namespace scifir
 
 			float get_r() const
 			{
-				return std::sqrt(std::pow(x,2) + std::pow(y,2) + std::pow(z,2));
+				return float(std::sqrt(std::pow(x,2) + std::pow(y,2) + std::pow(z,2)));
 			}
 
 			angle get_spherical_phi() const
@@ -815,7 +815,7 @@ namespace scifir
 
 			float distance_to_origin() const
 			{
-				return std::sqrt(std::pow(x,2) + std::pow(y,2) + std::pow(z,2));
+				return float(std::sqrt(std::pow(x,2) + std::pow(y,2) + std::pow(z,2)));
 			}
 
 			string display_cylindrical() const
