+++ coordinates/coordinates_nd.cpp
@@ -9,9 +9,9 @@ namespace scifir
 		if (x1.get_nd() == x2.get_nd())
 		{
 			float x_length = 0;
-			for (int i = 0; i < x1.values.size(); i++)
+			for (unsigned int i = 0; i < x1.values.size(); i++)
 			{
-				x_length += std::pow(x1.values[i] - x2.values[i],2);
+				x_length += float(std::pow(x1.values[i] - x2.values[i],2));
 			}
 			return std::sqrt(x_length);
 		}
@@ -26,9 +26,9 @@ namespace scifir
 		if (x1.get_nd() == x2.get_nd())
 		{
 			float x_length = 0;
-			for (int i = 0; i < x1.values.size(); i++)
+			for (unsigned int i = 0; i < x1.values.size(); i++)
 			{
-				x_length += std::pow(x1.values[i] - x2.values[i],2);
+				x_length += float(std::pow(x1.values[i] - x2.values[i],2));
 			}
 			return std::sqrt(x_length);
 		}
@@ -43,9 +43,9 @@ namespace scifir
 		if (x1.get_nd() == x2.get_nd())
 		{
 			float x_length = 0;
-			for (int i = 0; i < x1.values.size(); i++)
+			for (unsigned int i = 0; i < x1.values.size(); i++)
 			{
-				x_length += std::pow(x1.values[i] - x2.values[i],2);
+				x_length += float(std::pow(x1.values[i] - x2.values[i],2));
 			}
 			return std::sqrt(x_length);
 		}
