+++ coordinates/coordinates_nd.hpp
@@ -399,7 +399,7 @@ namespace scifir
 
 			T distance_to_origin() const
 			{
-				scalar_unit x_T = scalar_unit(0,"m2");
+				scalar_unit x_T = scalar_unit(0.0f,"m2");
 				for (int i = 0; i < values.size(); i++)
 				{
 					x_T += scifir::pow(values[i],2);
@@ -579,21 +579,21 @@ namespace scifir
 				return *this;
 			}
 
-			bool is_nd(int i) const
+			bool is_nd(unsigned int i) const
 			{
 				return values.size() == i;
 			}
 
 			int get_nd() const
 			{
-				return values.size();
+				return int(values.size());
 			}
 
 			float get_p() const
 			{
 				if (get_nd() == 2 or get_nd() == 3)
 				{
-					return std::sqrt(std::pow(values[0],2) + std::pow(values[1],2));
+					return float(std::sqrt(std::pow(values[0],2) + std::pow(values[1],2)));
 				}
 				else
 				{
@@ -617,7 +617,7 @@ namespace scifir
 			{
 				if (get_nd() == 2 or get_nd() == 3)
 				{
-					return std::sqrt(std::pow(values[0],2) + std::pow(values[1],2) + std::pow(values[2],2));
+					return float(std::sqrt(std::pow(values[0],2) + std::pow(values[1],2) + std::pow(values[2],2)));
 				}
 				else
 				{
@@ -823,9 +823,9 @@ namespace scifir
 			float distance_to_origin() const
 			{
 				float x_T = 0;
-				for (int i = 0; i < values.size(); i++)
+				for (unsigned int i = 0; i < values.size(); i++)
 				{
-					x_T += std::pow(values[i],2);
+					x_T += float(std::pow(values[i],2));
 				}
 				return std::sqrt(x_T);
 			}
@@ -883,7 +883,7 @@ namespace scifir
 	{
 		if (x1.get_nd() == x2.get_nd())
 		{
-			scalar_unit x_length = scalar_unit(0,"m2");
+			scalar_unit x_length = scalar_unit(0.0f,"m2");
 			for (int i = 0; i < x1.values.size(); i++)
 			{
 				x_length += scifir::pow(x1.values[i] - x2.values[i],2);
@@ -903,7 +903,7 @@ namespace scifir
 	{
 		if (x1.get_nd() == x2.get_nd())
 		{
-			scalar_unit x_length = scalar_unit(0,"m2");
+			scalar_unit x_length = scalar_unit(0.0f,"m2");
 			for (int i = 0; i < x1.values.size(); i++)
 			{
 				x_length += scifir::pow(x1.values[i] - x2.values[i],2);
@@ -923,7 +923,7 @@ namespace scifir
 	{
 		if (x1.get_nd() == x2.get_nd())
 		{
-			scalar_unit x_length = scalar_unit(0,"m2");
+			scalar_unit x_length = scalar_unit(0.0f,"m2");
 			for (int i = 0; i < x1.values.size(); i++)
 			{
 				x_length += scifir::pow(x1.values[i] - x2.values[i],2);
