+++ coordinates/coordinates_ndr.cpp
@@ -12,9 +12,9 @@ namespace scifir
 		if (x1.get_nd() == x2.get_nd())
 		{
 			float x_length = 0;
-			for (int i = 0; i < x1.get_values().size(); i++)
+			for (unsigned int i = 0; i < x1.get_values().size(); i++)
 			{
-				x_length += std::pow(x1.get_value(i) - x2.get_value(i),2);
+				x_length += float(std::pow(x1.get_value(i) - x2.get_value(i),2));
 			}
 			return std::sqrt(x_length);
 		}
@@ -29,9 +29,9 @@ namespace scifir
 		if (x1.get_nd() == x2.get_nd())
 		{
 			float x_length = 0;
-			for (int i = 0; i < x1.values.size(); i++)
+			for (unsigned int i = 0; i < x1.values.size(); i++)
 			{
-				x_length += std::pow(x1.values[i] - x2.values[i],2);
+				x_length += float(std::pow(x1.values[i] - x2.values[i],2));
 			}
 			return std::sqrt(x_length);
 		}
@@ -46,9 +46,9 @@ namespace scifir
 		if (x1.get_nd() == x2.get_nd())
 		{
 			float x_length = 0;
-			for (int i = 0; i < x1.values.size(); i++)
+			for (unsigned int i = 0; i < x1.values.size(); i++)
 			{
-				x_length += std::pow(x1.values[i] - x2.values[i],2);
+				x_length += float(std::pow(x1.values[i] - x2.values[i],2));
 			}
 			return std::sqrt(x_length);
 		}
@@ -63,9 +63,9 @@ namespace scifir
 		if (x1.get_nd() == x2.get_nd())
 		{
 			float x_length = 0;
-			for (int i = 0; i < x1.values.size(); i++)
+			for (unsigned int i = 0; i < x1.values.size(); i++)
 			{
-				x_length += std::pow(x1.values[i] - x2.values[i],2);
+				x_length += float(std::pow(x1.values[i] - x2.values[i],2));
 			}
 			return std::sqrt(x_length);
 		}
@@ -80,9 +80,9 @@ namespace scifir
 		if (x1.get_nd() == x2.get_nd())
 		{
 			float x_length = 0;
-			for (int i = 0; i < x1.values.size(); i++)
+			for (unsigned int i = 0; i < x1.values.size(); i++)
 			{
-				x_length += std::pow(x1.values[i] - x2.values[i],2);
+				x_length += float(std::pow(x1.values[i] - x2.values[i],2));
 			}
 			return std::sqrt(x_length);
 		}
