+++ topology/point_2d.cpp
@@ -18,6 +18,6 @@ namespace scifir
 
 	float distance(const point_2d<float>& x1,const point_2d<float>& x2)
 	{
-		return std::sqrt(std::pow(x1.x - x2.x,2) + std::pow(x1.y - x2.y,2));
+		return float(std::sqrt(std::pow(x1.x - x2.x,2) + std::pow(x1.y - x2.y,2)));
 	}
 }
