+++ topology/point_3d.cpp
@@ -19,6 +19,6 @@ namespace scifir
 
 	float distance(const point_3d<float>& x1,const point_3d<float>& x2)
 	{
-		return std::sqrt(std::pow(x1.x - x2.x,2) + std::pow(x1.y - x2.y,2) + std::pow(x1.z - x2.z,2));
+		return float(std::sqrt(std::pow(x1.x - x2.x,2) + std::pow(x1.y - x2.y,2) + std::pow(x1.z - x2.z,2)));
 	}
 }
