+++ topology/point_nd.hpp
@@ -398,7 +398,7 @@ namespace scifir
 
 			T distance_to_origin() const
 			{
-				scalar_unit x_T = scalar_unit(0,"m2");
+				scalar_unit x_T = scalar_unit(0.0f,"m2");
 				for (int i = 0; i < values.size(); i++)
 				{
 					x_T += scifir::pow(values[i],2);
@@ -573,21 +573,21 @@ namespace scifir
 
 			point_nd<float>& operator=(const coordinates_nd<float>&);
 
-			bool is_nd(int i) const
+			bool is_nd(unsigned int i) const
 			{
 				return values.size() == i;
 			}
 
 			int get_nd() const
 			{
-				return values.size();
+				return int(values.size());
 			}
 
 			float get_p() const
 			{
 				if (get_nd() == 2 or get_nd() == 3)
 				{
-					return std::sqrt(std::pow(values[0],2) + std::pow(values[1],2));
+					return float(std::sqrt(std::pow(values[0],2) + std::pow(values[1],2)));
 				}
 				else
 				{
@@ -611,7 +611,7 @@ namespace scifir
 			{
 				if (get_nd() == 2 or get_nd() == 3)
 				{
-					return std::sqrt(std::pow(values[0],2) + std::pow(values[1],2) + std::pow(values[2],2));
+					return float(std::sqrt(std::pow(values[0],2) + std::pow(values[1],2) + std::pow(values[2],2)));
 				}
 				else
 				{
@@ -817,9 +817,9 @@ namespace scifir
 			float distance_to_origin() const
 			{
 				float x_T = 0;
-				for (int i = 0; i < values.size(); i++)
+				for (unsigned int i = 0; i < values.size(); i++)
 				{
-					x_T += std::pow(values[i],2);
+					x_T += float(std::pow(values[i],2));
 				}
 				return std::sqrt(x_T);
 			}
